Union test-main

+++ include/turtletorque3/turtletorque3_motor_driver.h
@@ -82,10 +82,10 @@ class TurtleTorque3MotorDriver
   bool setTorque(bool onoff);
   bool getTorque();
   bool readEncoder(int32_t &left_value, int32_t &right_value);
-  bool readCurrent(int32_t &left_value, int32_t &right_value);
+  bool readCurrent(uint16_t &left_value, uint16_t &right_value);
   bool writeVelocity(int64_t left_value, int64_t right_value);
-  bool writeCurrent(int32_t left_value, int32_t right_value);                   //No sabemos si es int64 o int32
-  bool controlMotor(int32_t left_value, int32_t right_value);
+  bool writeCurrent(int left_value, int right_value);                   //No sabemos si es int64 o int32
+  bool controlMotor(int left_value, int right_value);
 
  private:
   uint32_t baudrate_;
